% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp_human}
\alias{emp_human}
\title{emp_human}
\format{
phyloseq object
}
\source{
https://www.nature.com/articles/nature24621
}
\usage{
data(emp_human)
}
\value{
Loads the data set in R.
}
\description{
The emp_human data set contains high-through taxonomic profiling data from
160 samples which is a small subet from the Earth Microbiome Project. The ASV-level counts
as would be expected for denoising algorithms.
#' NOTE: These are only test datasets.
}
\references{
Thompson LR et al. (2017) A communal catalogue reveals Earth’s multiscale microbial diversity
\url{https://www.nature.com/articles/nature24621}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{data}
