% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_microbiome_data.R
\name{list_microbiome_data}
\alias{list_microbiome_data}
\title{List of available datasets}
\usage{
list_microbiome_data(printtab = TRUE)
}
\arguments{
\item{printtab}{Print in console or not, defaut is TRUE and will print output. It is duplicate from microbiomeutilities pkg.}
}
\description{
Data are used from Duvallet et al 2017 https://www.nature.com/articles/s41467-017-01973-8.pdf?origin=ppub.
}
\details{
Data for practice. The source file for these data is the microbiomedatarepo https://github.com/microsud/microbiomedatarepo.
}
\examples{
\dontrun{
library(jeevanuDB)

df <- list_microbiome_data(printtab = FALSE)
}

}
\keyword{utilities}
