% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moving_pictures}
\alias{moving_pictures}
\title{moving_pictures}
\format{
phyloseq object
}
\source{
https://www.nature.com/articles/nature24621
}
\usage{
data(moving_pictures)
}
\value{
Loads the data set in R.
}
\description{
The moving_pictures data set contains high-through taxonomic profiling data from
1966 samples from two F4 (saliva n= 135, skin n= 268, stool n= 131) and M3 (saliva n= 373, skin n= 723, stool n= 336)
subjects which is a subet from the Earth Microbiome Project and samples from
the moving pictures study by . The ASV-level counts as would be expected for denoising algorithms.
NOTE: These are only test datasets.
}
\references{
\itemize{
\item{}{Thompson, Luke R., et al. "A communal catalogue reveals Earth’s multiscale microbial diversity." Nature 551, no. 7681 (2017): 457-463.}
\item{}{Caporaso, J. Gregory, et al. "Moving pictures of the human microbiome." Genome biology 12, no. 5 (2011): R50.}
\item{}{To cite the package, see citation('jeevanuDB')}
}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{data}
